requirements.txt
Django==4.2
crispy-bootstrap5==0.7
django-allauth==0.54.0
django-crispy-forms==2.0
django-countries==7.5.1
django-filter==23.2
django-tables2==2.6
flake8
python-dotenv==1.0.0
PyJWT==2.4.0
requests
